server/requirements.txt
attrs==22.2.0
cassandra-driver==3.25.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
geomet==0.2.1.post1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
python-dotenv==0.21.0
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.11.0
